moved player and wall classes to the entities directory
