+++ Devoid/Entities/Wall.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 
-namespace Devoid
+namespace Devoid.Entities
 {
     public class Wall
     {
